% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outputs.R
\name{outputs-extraction}
\alias{outputs-extraction}
\alias{ql_outputs}
\alias{outputs}
\alias{results}
\alias{errors}
\alias{warnings}
\alias{messages}
\alias{ql_results}
\alias{ql_errors}
\alias{ql_warnings}
\alias{ql_messages}
\title{Extract Common Output Types from an Evaluate Object}
\usage{
ql_outputs(class, field)

ql_results()

ql_errors()

ql_warnings()

ql_messages()
}
\arguments{
\item{class}{Character. The class of objects to extract from the \code{evaluate} output.}

\item{field}{Character. The field within the object to extract (e.g., \code{"value"} or \code{"message"}).}
}
\value{
A list containing the extracted fields from the specified class of outputs.
}
\description{
These functions facilitate the extraction of specific output types, such as results, errors, warnings,
or messages, from an \code{evaluate} object. The outputs are filtered based on their class and a specified
field, making it easier to analyse the results of code evaluation in Quarto Live exercise grading.
}
\details{
The functions operate on an \code{.evaluate_result} object, which stores outputs from code execution
(e.g., in Quarto Live exercises). They are designed to extract specific elements such as:
\itemize{
\item \strong{Results}: Extracts the \code{value} field from objects of class \code{"results"}.
\item \strong{Errors}: Extracts the \code{message} field from objects of class \code{"error"}.
\item \strong{Warnings}: Extracts the \code{message} field from objects of class \code{"warning"}.
\item \strong{Messages}: Extracts the \code{message} field from objects of class \code{"message"}.
}
}
\examples{
\dontshow{if (requireNamespace("evaluate", quietly = TRUE)) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}

.evaluate_result <- evaluate::evaluate(
'print(rnorm(10))
sample$x
log(-1)
message("Hello world")', output_handler = ql_output_handler)

# Extract results
ql_results()

# Extract errors
ql_errors()

# Extract warnings
ql_warnings()

# Extract messages
ql_messages()
\dontshow{\}) # examplesIf}
}
