% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outputs.R
\docType{data}
\name{ql_output_handler}
\alias{ql_output_handler}
\title{This mimics the behaviour of quarto-live evaluate results}
\format{
An object of class \code{output_handler} of length 8.
}
\usage{
ql_output_handler
}
\description{
This mimics the behaviour of quarto-live evaluate results
}
\keyword{internal}
