% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qlcheckr-package.R
\docType{package}
\name{qlcheckr-package}
\alias{qlcheckr}
\alias{qlcheckr-package}
\title{qlcheckr: Helpful Code Checking for Quarto Live}
\description{
Provides a framework for interactive code checking in Quarto Live. The package enables users to define logical tests for code correctness, delivers custom feedback messages, and includes debugging capabilities to display evaluation details and error hints. Designed to enhance teaching, learning, and automated code assessment workflows.
}
\author{
\strong{Maintainer}: Mitchell O'Hara-Wild \email{mail@mitchelloharawild.com} (\href{https://orcid.org/0000-0001-6729-7695}{ORCID})

}
\keyword{internal}
