% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{search_ast}
\alias{search_ast}
\title{Search Abstract Syntax Tree (AST) for Specific Patterns}
\usage{
search_ast(.code, .fn = NULL, ..., .expr = NULL)
}
\arguments{
\item{.code}{A parsed R expression or a call object representing the user code to analyse.
Typically obtained using \code{\link[=parse]{parse()}} or similar methods.}

\item{.fn}{Character or NULL. The name of the function to search for in the code.
If \code{NULL}, the function usage is not checked.}

\item{...}{Additional arguments to search for in function calls.
These may include named or unnamed arguments.}

\item{.expr}{A quoted or parsed R expression to search for within the code.
If \code{NULL}, expressions are not checked.}
}
\value{
A logical value indicating whether the specified patterns were found
in the provided code (\code{TRUE}) or not (\code{FALSE}).
}
\description{
Analyses parsed user code to identify the presence of specific patterns in the Abstract Syntax Tree (AST),
such as the usage of functions, arguments, or expressions.
}
\details{
The function operates on the AST representation of the provided code, enabling
precise pattern matching. Use \code{.fn} to check for the presence of a specific
function, \code{...} to search for arguments, and \code{.expr} for custom expression
matching. Any combination of these can be used to tailor the search criteria.
}
\examples{
# Example: Search for a specific function name
search_ast(quote(mean(x)), .fn = mean)

# Example: Search for a specific argument
search_ast(quote(mean(x, na.rm = TRUE)), na.rm = TRUE)
search_ast(quote(mean(x, na.rm = TRUE)), na.rm = FALSE)

# Example: Search for an expression
search_ast(quote(mean(x + y)), .expr = x + y)
search_ast(quote(mean(x + y)), .expr = mean(x + y))
search_ast(quote(mean(x + y)), .expr = log(x + y))

}
